Refactor of app.py completed Moved a copy of app.py to archive for further reference
